requirements.txt
bs4
user_agent
readability-lxml
textstat
langdetect